feat(chat): improved text input/send money experience
Make chat/text input a first class citizen in the messaging experience
