+++ apps/flipcash/features/messenger/src/main/kotlin/com/flipcash/app/messenger/internal/screens/components/ChatTopBar.kt
@@ -1,0 +1,80 @@
+package com.flipcash.app.messenger.internal.screens.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.requiredSize
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+import com.flipcash.app.contacts.ui.ContactAvatar
+import com.flipcash.app.core.contacts.DeviceContact
+import com.getcode.navigation.core.CodeNavigator
+import com.getcode.theme.CodeTheme
+import com.getcode.ui.components.AppBarDefaults
+import com.getcode.ui.components.AppBarWithTitle
+import com.getcode.ui.core.measured
+
+@Composable
+internal fun ChatTopBar(
+    navigator: CodeNavigator,
+    chattingWith: DeviceContact?,
+) {
+    var titleHeight by remember { mutableStateOf(0.dp) }
+    val bgColor = CodeTheme.colors.background
+    Box {
+        Box(
+            modifier = Modifier
+                .fillMaxWidth()
+                .height(titleHeight + 24.dp)
+                .background(
+                    Brush.verticalGradient(
+                        colorStops = arrayOf(
+                            0f to bgColor,
+                            1f to Color.Transparent,
+                        )
+                    )
+                )
+        )
+        AppBarWithTitle(
+            modifier = Modifier.measured { titleHeight = it.height },
+            leftIcon = {
+                AppBarDefaults.UpNavigation { navigator.pop() }
+            },
+            title = {
+                Row(
+                    modifier = Modifier.fillMaxWidth(),
+                    verticalAlignment = Alignment.CenterVertically,
+                    horizontalArrangement = Arrangement.spacedBy(CodeTheme.dimens.grid.x2),
+                ) {
+                    ContactAvatar(
+                        contact = chattingWith,
+                        modifier = Modifier
+                            .requiredSize(CodeTheme.dimens.staticGrid.x8)
+                            .clip(CircleShape),
+                    )
+
+                    Text(
+                        modifier = Modifier.weight(1f),
+                        text = chattingWith?.displayName.orEmpty(),
+                        style = CodeTheme.typography.textMedium,
+                        color = CodeTheme.colors.textMain,
+                    )
+                }
+            }
+        )
+    }
+}
